Improve Image GET
Now returns just a path rather than the base64 image. Should use less bandwidth since base64 is roughly 4/3 as large and it also just makes more sense. We're now returning the full path from the image directory and also the relative path from the image directory. The full path is to load the image client-side and the relative is for when the path is sent to the server in the chart config it is independent of the server setup, which I believe is the correct decision.