Relative GET Image path
Having an image path start with "/" is needed for the front end img.src but an image path without that is needed for POSTing the image. Not sure how to ameliorate this other than having two IMAGE_DIRs for each scenario. This may just be a DJANGO problem that I haven't figured out.